WIP: Add more shared libraries
When using shared libaries, the visibility=hidden
is used for shared libraries

Adding Core/Mesh and Core/SpatialObjects to shared
libaries.

The intent of this patch is to allow gcc and clang
to emulate the same symbol exporting behavior as
windows so that external symbols can be tested
more conveniently.

There are benefits to having symbols hidden, including
improved startup time of shared libraries, and smaller
libraries.

Change-Id: Ied324c415b4cce8bbab6239f46efaf73597f4b95